Wrap updateDisplayAllBtn()
